Fix false-positive CVC errors in credit card donation flow.

+++ app/src/main/java/org/thoughtcrime/securesms/components/settings/app/subscription/donate/card/CreditCardFormState.kt
@@ -20,7 +20,7 @@ data class CreditCardFormState(
       number,
       expiration.month.toInt(),
       expiration.year.toInt(),
-      code.toInt()
+      code
     )
   }
 }
